Update BuildSys.R
Fix bugs on MacOS and generalise the debug architecture. MacOS version of R GUI is a pain in the arse cos it has no consistency of behaviour with Windows or Linux R behaviour. R command line version is but llbd on mac has no support for external console debugging so debugging the command line version is not workable.