Update BuildSys.R
Tentative fix for the Ripley problem. Need to test thoroughly and if it appears to work then remove the commented lines and make permanent. Might work on R hub but still fail in CRAN though (if he's using static R library). Might need a more sophisticated approach that takes that into account. Should see if I can figure out how that will be presented and add as needed.